% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_files.R
\name{move_files}
\alias{move_files}
\title{Move or Rename Files}
\usage{
move_files(old_dir, new_dir, wildcard = NULL)
}
\arguments{
\item{old_dir, new_dir}{File path ending in a slash}

\item{wildcard}{Character pattern with glob2rx syntax with which to select files}
}
\description{
Move or Rename Files
}
